backend/requirements.txt
flask==2.0.1
flask-cors==3.0.10
opencv-python==4.8.1.78
numpy==1.26.3
pillow==10.2.0
werkzeug==2.3.7
rembg==2.0.65
imutils==0.5.4
pandas==2.1.4
openpyxl==3.1.2 